test(mysql): align SAB-514 test coverage
